所有的label都转成float32

+++ mmseg/models/losses/eps_loss.py
@@ -324,7 +324,8 @@ def eps_wsss_loss(pred,
     
     b,c,h,w=pred.size()
     label = torch.sum(label.view(b, c, -1),dim=-1)
-    label[label>0]=1
+    label[label>0]=1.
+    label=label.float()
     # label = label.unsqueeze(-1).unsqueeze(-1)
     # pdb.set_trace()
     pred_=torch.roll(target[-1], shifts=-1, dims=1)#[bg,classes ch,..] [bs,c,h,w]
